pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["PATH242 <mennashaheen2357@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = {version = "^0.29.0", extras = ["standard"]}
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
streamlit-option-menu = "^0.3.12"
streamlit = "^1.33.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.1.2"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
mypy = "^1.10.0"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240327"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
